buyer can see all his oders functionality

+++ p1/src/main/java/com/revshop/p1/repository/SellerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 import com.revshop.p1.entity.Seller;
 @Repository
-public interface SellerRepository extends JpaRepository<Seller, Integer>{
+public interface SellerRepository extends JpaRepository<Seller, Long>{
 	Seller findByEmailAndPassword(String email,String password);
 	Optional<Seller> findByEmail(String email);
 }
